Minor changes and such.

+++ src/main/java/com/worldcretornica/plotme/PlotRoadPopulator.java
@@ -1,10 +1,10 @@
 package com.worldcretornica.plotme;
 
-import java.util.Random;
-
 import org.bukkit.Chunk;
 import org.bukkit.World;
 import org.bukkit.generator.BlockPopulator;
+
+import java.util.Random;
 
 public class PlotRoadPopulator extends BlockPopulator
 {
@@ -85,15 +85,15 @@ public class PlotRoadPopulator extends BlockPopulator
 		
 		if(pathsize % 2 == 1)
 		{
-			n1 = Math.ceil(((double)pathsize)/2) - 2;
-			n2 = Math.ceil(((double)pathsize)/2) - 1;
-			n3 = Math.ceil(((double)pathsize)/2);
+			n1 = Math.ceil(pathsize /2) - 2;
+			n2 = Math.ceil(pathsize /2) - 1;
+			n3 = Math.ceil(pathsize /2);
 		}
 		else
 		{
-			n1 = Math.floor(((double)pathsize)/2) - 2;
-			n2 = Math.floor(((double)pathsize)/2) - 1;
-			n3 = Math.floor(((double)pathsize)/2);
+			n1 = Math.floor(pathsize /2) - 2;
+			n2 = Math.floor(pathsize /2) - 1;
+			n3 = Math.floor(pathsize /2);
 		}
 		
 		if(pathsize % 2 == 1)
